+++ src/execute/runtimes/node/controllable_worker_thread.mjs
@@ -1,34 +1,10 @@
 import { parentPort } from "node:worker_threads"
 import { uneval } from "@jsenv/uneval"
-import { startObservingPerformances } from "./node_execution_performance.js"
+
+import { executeUsingDynamicImport } from "./execute_using_dynamic_import.js"
 
 const ACTIONS_AVAILABLE = {
-  "execute-using-dynamic-import": async ({ fileUrl, collectPerformance }) => {
-    const getNamespace = async () => {
-      const namespace = await import(fileUrl)
-      const namespaceResolved = {}
-      await Promise.all([
-        ...Object.keys(namespace).map(async (key) => {
-          const value = await namespace[key]
-          namespaceResolved[key] = value
-        }),
-      ])
-      return namespaceResolved
-    }
-    if (collectPerformance) {
-      const getPerformance = startObservingPerformances()
-      const namespace = await getNamespace()
-      const performance = await getPerformance()
-      return {
-        namespace,
-        performance,
-      }
-    }
-    const namespace = await getNamespace()
-    return {
-      namespace,
-    }
-  },
+  "execute-using-dynamic-import": executeUsingDynamicImport,
 }
 const ACTION_REQUEST_EVENT_NAME = "action"
 const ACTION_RESPONSE_EVENT_NAME = "action-result"