+++ src/execute/runtimes/node/node_child_process.js
@@ -36,7 +36,8 @@ nodeChildProcess.run = async ({
   stopSignal,
   onConsole,
 
-  collectCoverage = false,
+  coverageEnabled = false,
+  coverageConfig,
   coverageMethodForNodeJs,
   collectPerformance,
 
@@ -55,7 +56,6 @@ nodeChildProcess.run = async ({
   }
   env = {
     ...env,
-    COVERAGE_ENABLED: collectCoverage,
     JSENV: true,
   }
   if (coverageMethodForNodeJs !== "NODE_V8_COVERAGE") {
@@ -193,8 +193,12 @@ nodeChildProcess.run = async ({
       data: {
         actionType: "execute-using-dynamic-import",
         actionParams: {
+          rootDirectoryUrl,
           fileUrl: new URL(fileRelativeUrl, rootDirectoryUrl).href,
           collectPerformance,
+          coverageEnabled,
+          coverageConfig,
+          coverageMethodForNodeJs,
           exitAfterAction: true,
         },
       },