+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Rename nodeProcess to nodeChildProcess
 - Add nodeWorkerThread
+- Rename executeTestPlan.coverage to executeTestPlan.coverageEnabled
+- When coverage is enabled, process.env.NODE_V8_COVERAGE must be set otherwise a warning is logged
 
 # 27.1.0